move entry functions into root module
