+++ src/lib.rs
@@ -15,8 +15,10 @@ mod syscall_mock;
 mod tracer;
 pub mod utils;
 
-use crate::syscall_mock::{executable_mock, run_against_protocols};
-use protocol::Protocols;
+use crate::protocol::{Protocol, Protocols};
+use crate::syscall_mock::test_result::{TestResult, TestResults};
+use crate::syscall_mock::{executable_mock, SyscallMock};
+use crate::tracer::Tracer;
 use std::io::Write;
 use std::path::{Path, PathBuf};
 
@@ -144,4 +146,113 @@ pub fn run_check_protocols(context: Context, script: &Path) -> R<(ExitCode, Stri
     let expected_protocols = Protocols::load(script)?;
     let results = run_against_protocols(context, script, expected_protocols)?;
     Ok((results.exitcode(), results.format_test_results()))
+}
+
+fn run_against_protocols(
+    context: Context,
+    program: &Path,
+    Protocols {
+        protocols,
+        unmocked_commands,
+        interpreter,
+    }: Protocols,
+) -> R<TestResults> {
+    Ok(TestResults(
+        protocols
+            .into_iter()
+            .map(|expected| {
+                run_against_protocol(
+                    context.clone(),
+                    &interpreter,
+                    program,
+                    expected,
+                    &unmocked_commands,
+                )
+            })
+            .collect::<R<Vec<TestResult>>>()?,
+    ))
+}
+
+pub fn run_against_protocol(
+    context: Context,
+    interpreter: &Option<Vec<u8>>,
+    program: &Path,
+    expected: Protocol,
+    unmocked_commands: &[Vec<u8>],
+) -> R<TestResult> {
+    let syscall_mock = Tracer::run_against_mock(
+        interpreter,
+        program,
+        expected.arguments.clone(),
+        expected.env.clone(),
+        |tracee_pid| SyscallMock::new(context, tracee_pid, expected, unmocked_commands),
+    )?;
+    Ok(syscall_mock.result)
+}
+
+#[cfg(test)]
+mod run_against_protocol {
+    use super::*;
+    use crate::protocol::command::Command;
+    use std::fs;
+    use std::os::unix::ffi::OsStrExt;
+    use test_utils::{trim_margin, TempFile};
+
+    #[test]
+    fn works_for_longer_file_names() -> R<()> {
+        let long_command = TempFile::new()?;
+        fs::copy("/bin/true", long_command.path())?;
+        let script = TempFile::write_temp_script(
+            trim_margin(&format!(
+                r##"
+                    |#!/usr/bin/env bash
+                    |{}
+                "##,
+                long_command.path().to_string_lossy()
+            ))?
+            .as_bytes(),
+        )?;
+        assert_eq!(
+            run_against_protocol(
+                Context::new_test_context(),
+                &None,
+                &script.path(),
+                Protocol::new(vec![protocol::Step {
+                    command: Command {
+                        executable: long_command.path().as_os_str().as_bytes().to_vec(),
+                        arguments: vec![],
+                    },
+                    stdout: vec![],
+                    exitcode: 0
+                }]),
+                &[]
+            )?,
+            TestResult::Pass
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn does_not_execute_the_commands() -> R<()> {
+        let testfile = TempFile::new()?;
+        let script = TempFile::write_temp_script(
+            trim_margin(&format!(
+                r##"
+                    |#!/usr/bin/env bash
+                    |touch {}
+                "##,
+                testfile.path().to_string_lossy()
+            ))?
+            .as_bytes(),
+        )?;
+        let _ = run_against_protocol(
+            Context::new_test_context(),
+            &None,
+            &script.path(),
+            Protocol::empty(),
+            &[],
+        )?;
+        assert!(!testfile.path().exists(), "touch was executed");
+        Ok(())
+    }
 }
